1st Commit - Overview
It currently works well in Gmail and Newer versions of Outlook across the web and mobile apps. However, it does not work as intended on older versions of Outlook. One of the older versions where it was tested and it did not work was the following:Microsoft® Outlook® for Microsoft 365 MSO (Version 2405 Build 16.0.17628.20006) 64-bit. The problem is that the background-image CSS property is being removed. The intended design that should be followed can be found with this commit as a png file.