Background Image Width Set To 100% Works However ...
Even though the width is no longer shorter than the section directly above and they are the same, it makes the entire email wider than it's intended/desired design. The issue is referenced in this link - https://stackoverflow.com/questions/23280715/full-width-backgrounds-on-outlook-07-10-13 You can see the visual difference through the attached images. This happened by switching from width:600px to width:1000