Footer Background Image Width Aligned With Body/Top Section's Width
I had to play around with various widths. Some were too short and too long. I'm concerned that even though the widths are aligned for each section on my email that it may not be for other team members. It's still unclear why 100% doesn't work without making the entire email wider. We just need it to fill the available space but I haven't been able to find such a solution yet.