Refer To Previous Commit - Footer Background Image Width Aligned With Body/Top Section's Width
The attached images highlight that the width is too large when the emails are double clicked which opens it in a new window which is wider because the normal elements on the left aren't shown. The email that isn't double clicked is shown in the image labelled 'Elements On The Left - Tabs And List of Emails - Shayn Hacker'. When it's double clicked and shown in a new window, the viewport of the screen is larger and hence the background image is larger because it's currently set to %. It seems that it needs to be a fixed width. Earlier testing had it set to 600 but it was slightly shorter on my Outlook, even shorter on Jerome's and even shorter on Shamari's. However, it's way too wide now on all of our devices. Interestingly, it isn't very wide when it's double clicked on emails where I didn't specify % but rather pixels. If I switch it from mso-width-percent:555 to width:100% the entire width of the email will/should become larger which makes it different from the desired/intended design.