OTP Email Template - Asset's Links Updated
It's been modified from the original version/folder provided by Dwayne. Previously, all of the assets (logo, social media icons) were referenced from the folder containing the email. However, the assets won't show when the email is sent because the email client won't have access to them, as only the file is sent, not the folder. Therefore, I had to replace the references to the assets inside the folder with external links that can be accessed anywhere. For example, I changed 'Sagicor-Logo.png' to 'https://www.sagicor.com/-/media/Email-Images/client-portal/Sagicor-Logo.png'. Please find image attachments that highlights the difference between the intended design and how it looks when an email is sent/received.