Insight - HTML Email Opened With Microsoft
I read that older versions of Outlook use the same rendering engine as Microsoft Word. It eventually dawned on me to try opening the HTML email with Microsoft Word to see if it would look the same as in the email environment. It's not exactly the same, but it's pretty similar. You can see that the background image actually shows, but if you use the original version of the email, it won't, since the original version doesn't have VML, which is what the Microsoft rendering engine recognizes. You can also see in the images attached that the width of the section containing most of the copy is shorter than the section containing the verification code. The next step is to see if edits made through Microsoft Word will reflect in the actual email environment and at the same time don't change how it looks in newer versions of Outlook, Gmail, and Yahoo.