Microsoft Edits Actually Worked However....
It has affected other parts of the email, particularly for the newer versions of Outlook and Gmail. The edit in question was making the part on the right that has the most copy/text white. Refer to previous commit for more context. The HTML file attached with this commit was changed due to saving the edits made with Microsoft Word. After this commit, I'll revert it to it's original state. I suppose this is more of a proof of concept that the closest we can get to a VML editor is Microsoft Word.