Refer To Previous Commit
I also forgot to mention that one of the other changes I made using Microsoft Word was reducing the amount of spacing between the copy for example there was a lot of spacing historically between 'Verify your email address' and 'Hi Rachel' and between the last sentence and the verification code. I did this by deleting the row through Word and the HTML tag equivalent would be <tr>. It doesn't seem to affect the newer Outlook or Gmail or at least in a major way.