Body & Footer Section's Width Are The Same
Dwayne noticed that the width of the body section was shorter than the footer section. I also noticed the same thing. Please refer to this ChatGPT chat screenshot that helped me in resolving the problem. There is no shareable link yet available for chats that have images. I had to use a conditional VML statement around the body section.Refer to Dwayne's screenshot to see how the widths aren't aligned. The next step is to reduce the excessive spacing between "Verify your email address" and "Hi Rachel," & between the last sentence in the body and "Verification Cody Copy"