Several Updates Required
It needs placeholder information to the right of Policy, Billed To, Date, Transaction ID and Amount. It has been tested in old/new Outlook and Gmail and works as intended based on the Figma file for the most part.  You can refer to the attached images to see the Figma file and how the email looks in different email clients.